Changed the layout of the start program.
You have to provide the path, where the data is stored as a program
argument. Example "C:\\Users\\Marius\\Desktop\\musicus_data\\" with the
quotation marks. The previous data folder has now be renamed to
resources and is to be used for resources the program needs.

A config file will soon be placed into that directory, to configure the
system from outside the program. This could help automation and
scalability of the program.