Implemented Stafflined detection by Orientation.
Also redesigned the staffline (Every pixel of the staffline is now saved
in the Staffline).
Also added a simple Staffline removal which takes the pixels of the
staffline and removes them without any further checking (tho single
pixels are filtered out to clean the picture a bit)

Start was changed to use the Orientation Detection and the simple
removal and results in VERY accurate results.